Fix test globbing on Windows
Windows doesn't support the inline evaluation `$(find..)` and also
doesn't support the find command. `nyc` can handle this itself, as can
it appears `_mocha` (i.e. this command works both without and with the
`--` given just before the globbing), so it doesn't seem the find
command is necessary.

This allows `npm test` to function on Windows using PowerShell. For some
reason the coverage report is missing data, but it misses data without
this change also.