Use a prepare script for transpiling instead of prepublish
From Travis logs:
As of npm@5, `prepublish` scripts will run only for `npm publish`.
(In npm@4 and previous versions, it also runs for `npm install`.)
See the deprecation note in `npm help scripts` for more information.

This was also causing discord-irc to fail to transpile when installed
using npm from source.

This modifies the script to run on the `prepare` hook, so it should
transpile when installed directly from source too.